requirements/dev.txt
#
# This file includes packages with requirements for local development.
#

django==3.1.2
djangorestframework==3.12.1
psycopg2-binary==2.8.6
django-environ==0.4.5
django-filter==2.4.0
django-seed==0.2.2
drf-yasg2==1.18.5
flake8==3.8.4